requirements.txt
streamlit
requests
torch
torchvision
transformers
pymongo
python-dotenv
cloudinary
easyocr
numpy
pillow
extra-streamlit-components
google-generativeai